# Merging slices from the labels APIs using more than 1 cpu and k-way